chore: move code from jax_distribution folder to mava folder and fix typing errors
