--- jax_distribution/requirements.txt
@@ -1,12 +1,0 @@
-distrax
-flax
-git+https://github.com/instadeepai/jumanji.git
-hydra-core==1.3.2
-jax<=0.4.13
-jaxlib<=0.4.13
-numpy
-omegaconf
-optax
-protobuf~=3.20
-sacred
-tensorboard_logger
