+++ mava/configs/rware_scenario/small-4ag.yaml
