+++ mava/configs/rware_scenario/tiny-2ag.yaml
