+++ mava/configs/rware_scenario/tiny-4ag-easy.yaml
