+++ mava/systems/rec_mappo_rware.py
@@ -11,7 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import copy
+import functools
 import os
 from logging import Logger as SacredLogger
 from os.path import abspath, dirname
@@ -31,28 +33,50 @@ from flax.core.frozen_dict import FrozenDict
 from flax.linen.initializers import constant, orthogonal
 from jumanji.env import Environment
 from jumanji.environments.routing.robot_warehouse.generator import RandomGenerator
-from jumanji.types import Observation
 from jumanji.wrappers import AutoResetWrapper
 from omegaconf import DictConfig, OmegaConf
 from optax._src.base import OptState
 from sacred import Experiment, observers, run, utils
 
-from jax_distribution.evaluator import evaluator_setup
-from jax_distribution.logger import logger_setup
-from jax_distribution.types import ExperimentOutput, LearnerState, PPOTransition
-from jax_distribution.utils.jax import merge_leading_dims
-from jax_distribution.utils.logger_tools import (
-    config_copy,
-    get_experiment_path,
-    get_logger,
-)
-from jax_distribution.utils.timing_utils import TimeIt
-from jax_distribution.wrappers.jumanji import (
+from mava.evaluator import evaluator_setup
+from mava.logger import logger_setup
+from mava.types import ExperimentOutput, PPOTransition, RNNLearnerState
+from mava.utils.logger_tools import config_copy, get_experiment_path, get_logger
+from mava.utils.timing_utils import TimeIt
+from mava.wrappers.jumanji import (
     AgentIDWrapper,
     LogWrapper,
     ObservationGlobalState,
     RwareMultiAgentWithGlobalStateWrapper,
 )
+
+
+class ScannedRNN(nn.Module):
+    @functools.partial(
+        nn.scan,
+        variable_broadcast="params",
+        in_axes=0,
+        out_axes=0,
+        split_rngs={"params": False},
+    )
+    @nn.compact
+    def __call__(self, carry: chex.Array, x: chex.Array) -> Tuple[chex.Array, chex.Array]:
+        """Applies the module."""
+        rnn_state = carry
+        ins, resets = x
+        rnn_state = jnp.where(
+            resets[:, jnp.newaxis],
+            self.initialize_carry(ins.shape[0], ins.shape[1]),
+            rnn_state,
+        )
+        new_rnn_state, y = nn.GRUCell()(rnn_state, ins)
+        return new_rnn_state, y
+
+    @staticmethod
+    def initialize_carry(batch_size: int, hidden_size: int) -> chex.Array:
+        """Initializes the carry state."""
+        # Use a dummy key since the default state init fn is just zeros.
+        return nn.GRUCell.initialize_carry(jax.random.PRNGKey(0), (batch_size,), hidden_size)
 
 
 class ActorCritic(nn.Module):
@@ -61,14 +85,25 @@ class ActorCritic(nn.Module):
     action_dim: Sequence[int]
 
     @nn.compact
-    def __call__(self, observation: Observation) -> Tuple[distrax.Categorical, chex.Array]:
+    def __call__(
+        self,
+        hiddens: Tuple[chex.Array, chex.Array],
+        local_observation: Tuple[chex.Array, chex.Array],
+    ) -> Tuple[Tuple[chex.Array, chex.Array], distrax.Categorical, chex.Array]:
         """Forward pass."""
-        x = observation.agents_view
+        observation, done = local_observation
+        policy_hidden_state, critic_hidden_state = hiddens
 
-        actor_output = nn.Dense(64, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0))(x)
-        actor_output = nn.relu(actor_output)
-        actor_output = nn.Dense(64, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0))(
-            actor_output
+        policy_embedding = nn.Dense(
+            128, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
+        )(observation.agents_view)
+        policy_embedding = nn.relu(policy_embedding)
+
+        policy_rnn_in = (policy_embedding, done)
+        policy_hidden_state, policy_embedding = ScannedRNN()(policy_hidden_state, policy_rnn_in)
+
+        actor_output = nn.Dense(128, kernel_init=orthogonal(2), bias_init=constant(0.0))(
+            policy_embedding
         )
         actor_output = nn.relu(actor_output)
         actor_output = nn.Dense(
@@ -80,21 +115,24 @@ class ActorCritic(nn.Module):
             actor_output,
             jnp.finfo(jnp.float32).min,
         )
-        actor_policy = distrax.Categorical(logits=masked_logits)
 
-        y = observation.global_state
+        pi = distrax.Categorical(logits=masked_logits)
 
-        critic_output = nn.Dense(64, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0))(y)
-        critic_output = nn.relu(critic_output)
-        critic_output = nn.Dense(64, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0))(
-            critic_output
-        )
-        critic_output = nn.relu(critic_output)
-        critic_output = nn.Dense(1, kernel_init=orthogonal(1.0), bias_init=constant(0.0))(
-            critic_output
-        )
+        critic_embedding = nn.Dense(
+            128, kernel_init=orthogonal(np.sqrt(2)), bias_init=constant(0.0)
+        )(observation.global_state)
+        critic_embedding = nn.relu(critic_embedding)
 
-        return actor_policy, jnp.squeeze(critic_output, axis=-1)
+        critic_rnn_in = (critic_embedding, done)
+        critic_hidden_state, critic_embedding = ScannedRNN()(critic_hidden_state, critic_rnn_in)
+
+        critic = nn.Dense(128, kernel_init=orthogonal(2), bias_init=constant(0.0))(critic_embedding)
+        critic = nn.relu(critic)
+        critic = nn.Dense(1, kernel_init=orthogonal(1.0), bias_init=constant(0.0))(critic)
+
+        hiddens = (policy_hidden_state, critic_hidden_state)
+
+        return hiddens, pi, jnp.squeeze(critic, axis=-1)
 
 
 def get_learner_fn(
@@ -102,7 +140,7 @@ def get_learner_fn(
 ) -> Callable:
     """Get the learner function."""
 
-    def _update_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, Tuple]:
+    def _update_step(learner_state: RNNLearnerState, _: Any) -> Tuple[RNNLearnerState, Tuple]:
         """A single update of the network.
 
         This function steps the environment and records the trajectory batch for
@@ -117,23 +155,53 @@ def get_learner_fn(
                 - rng (PRNGKey): The random number generator state.
                 - env_state (State): The environment state.
                 - last_timestep (TimeStep): The last timestep in the current trajectory.
+                - last_done (bool): Whether the last timestep was a terminal state.
+                - hstates (chex.Array): The hidden state of the RNN.
             _ (Any): The current metrics info.
         """
 
-        def _env_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, PPOTransition]:
+        def _env_step(
+            learner_state: RNNLearnerState, _: Any
+        ) -> Tuple[RNNLearnerState, PPOTransition]:
             """Step the environment."""
-            params, opt_state, rng, env_state, last_timestep = learner_state
+            (
+                params,
+                opt_state,
+                rng,
+                env_state,
+                last_timestep,
+                last_done,
+                hstates,
+            ) = learner_state
 
-            # SELECT ACTION
             rng, policy_rng = jax.random.split(rng)
-            actor_policy, value = apply_fn(params, last_timestep.observation)
+
+            # Add a batch dimension to the observation.
+            batched_observation = jax.tree_util.tree_map(
+                lambda x: x[jnp.newaxis, :], last_timestep.observation
+            )
+            ac_in = (
+                batched_observation,
+                last_done[:, 0][jnp.newaxis, :],
+            )
+
+            # Run the network.
+            hstates, actor_policy, value = apply_fn(params, hstates, ac_in)
+
+            # Sample action from the policy and squeeze out the batch dimension.
             action = actor_policy.sample(seed=policy_rng)
             log_prob = actor_policy.log_prob(action)
 
-            # STEP ENVIRONMENT
+            action, log_prob, value = (
+                action.squeeze(0),
+                log_prob.squeeze(0),
+                value.squeeze(0),
+            )
+
+            # Step the environment.
             env_state, timestep = jax.vmap(env.step, in_axes=(0, 0))(env_state, action)
 
-            # LOG EPISODE METRICS
+            # log episode return and length
             done, reward = jax.tree_util.tree_map(
                 lambda x: jnp.repeat(x, config["num_agents"]).reshape(config["num_envs"], -1),
                 (timestep.last(), timestep.reward),
@@ -146,8 +214,13 @@ def get_learner_fn(
             transition = PPOTransition(
                 done, action, value, reward, log_prob, last_timestep.observation, info
             )
-            learner_state = LearnerState(params, opt_state, rng, env_state, timestep)
+            learner_state = RNNLearnerState(
+                params, opt_state, rng, env_state, timestep, done, hstates
+            )
             return learner_state, transition
+
+        # INITIALISE RNN STATE
+        initial_hstates = learner_state.hstates
 
         # STEP ENVIRONMENT FOR ROLLOUT LENGTH
         learner_state, traj_batch = jax.lax.scan(
@@ -155,8 +228,30 @@ def get_learner_fn(
         )
 
         # CALCULATE ADVANTAGE
-        params, opt_state, rng, env_state, last_timestep = learner_state
-        _, last_val = apply_fn(params, last_timestep.observation)
+        (
+            params,
+            opt_state,
+            rng,
+            env_state,
+            last_timestep,
+            last_done,
+            hstates,
+        ) = learner_state
+
+        # Add a batch dimension to the observation.
+        batched_last_observation = jax.tree_util.tree_map(
+            lambda x: x[jnp.newaxis, :], last_timestep.observation
+        )
+        ac_in = (
+            batched_last_observation,
+            last_done[:, 0][jnp.newaxis, :],
+        )
+
+        # Run the network.
+        _, _, last_val = apply_fn(params, hstates, ac_in)
+        # Squeeze out the batch dimension and mask out the value of terminal states.
+        last_val = last_val.squeeze(0)
+        last_val = jnp.where(last_done, jnp.zeros_like(last_val), last_val)
 
         def _calculate_gae(
             traj_batch: PPOTransition, last_val: chex.Array
@@ -192,7 +287,13 @@ def get_learner_fn(
             def _update_minibatch(train_state: Tuple, batch_info: Tuple) -> Tuple:
                 """Update the network for a single minibatch."""
                 params, opt_state = train_state
-                traj_batch, advantages, targets = batch_info
+                (
+                    init_policy_hstate,
+                    init_critic_hstate,
+                    traj_batch,
+                    advantages,
+                    targets,
+                ) = batch_info
 
                 def _loss_fn(
                     params: FrozenDict,
@@ -203,7 +304,14 @@ def get_learner_fn(
                 ) -> Tuple:
                     """Calculate the loss."""
                     # RERUN NETWORK
-                    actor_policy, value = apply_fn(params, traj_batch.obs)
+                    _, actor_policy, value = apply_fn(
+                        params,
+                        (init_policy_hstate.squeeze(0), init_critic_hstate.squeeze(0)),
+                        (
+                            traj_batch.obs,
+                            traj_batch.done[:, :, 0],
+                        ),
+                    )
                     log_prob = actor_policy.log_prob(traj_batch.action)
 
                     # CALCULATE VALUE LOSS
@@ -241,7 +349,7 @@ def get_learner_fn(
                 # Compute the parallel mean (pmean) over the batch.
                 # This calculation is inspired by the Anakin architecture demo notebook.
                 # available at https://tinyurl.com/26tdzs5x
-                # This pmean could be a regular mean as the batch axis is on the same device.
+                # This pmean could be a regular mean as the batch axis is on all devices.
                 grads, loss_info = jax.lax.pmean((grads, loss_info), axis_name="batch")
                 # pmean over devices.
                 grads, loss_info = jax.lax.pmean((grads, loss_info), axis_name="device")
@@ -251,43 +359,82 @@ def get_learner_fn(
 
                 return (new_params, new_opt_state), loss_info
 
-            params, opt_state, traj_batch, advantages, targets, rng = update_state
+            (
+                params,
+                opt_state,
+                init_hstates,
+                traj_batch,
+                advantages,
+                targets,
+                rng,
+            ) = update_state
+            init_policy_hstate, init_critic_hstate = init_hstates
             rng, shuffle_rng = jax.random.split(rng)
 
             # SHUFFLE MINIBATCHES
-            batch_size = config["rollout_length"] * config["num_envs"]
-            permutation = jax.random.permutation(shuffle_rng, batch_size)
-            batch = (traj_batch, advantages, targets)
-            batch = jax.tree_util.tree_map(lambda x: merge_leading_dims(x, 2), batch)
-            shuffled_batch = jax.tree_util.tree_map(
-                lambda x: jnp.take(x, permutation, axis=0), batch
+            permutation = jax.random.permutation(shuffle_rng, config["num_envs"])
+            batch = (
+                init_policy_hstate,
+                init_critic_hstate,
+                traj_batch,
+                advantages,
+                targets,
             )
-            minibatches = jax.tree_util.tree_map(
-                lambda x: jnp.reshape(x, [config["num_minibatches"], -1] + list(x.shape[1:])),
+            shuffled_batch = jax.tree_util.tree_map(
+                lambda x: jnp.take(x, permutation, axis=1), batch
+            )
+            reshaped_batch = jax.tree_util.tree_map(
+                lambda x: jnp.reshape(x, (x.shape[0], config["num_minibatches"], -1, *x.shape[2:])),
                 shuffled_batch,
             )
+            minibatches = jax.tree_util.tree_map(lambda x: jnp.swapaxes(x, 1, 0), reshaped_batch)
 
             # UPDATE MINIBATCHES
             (params, opt_state), loss_info = jax.lax.scan(
                 _update_minibatch, (params, opt_state), minibatches
             )
 
-            update_state = (params, opt_state, traj_batch, advantages, targets, rng)
+            update_state = (
+                params,
+                opt_state,
+                init_hstates,
+                traj_batch,
+                advantages,
+                targets,
+                rng,
+            )
             return update_state, loss_info
 
-        update_state = (params, opt_state, traj_batch, advantages, targets, rng)
+        init_hstates = jax.tree_util.tree_map(lambda x: x[None, :], initial_hstates)
+        update_state = (
+            params,
+            opt_state,
+            init_hstates,
+            traj_batch,
+            advantages,
+            targets,
+            rng,
+        )
 
         # UPDATE EPOCHS
         update_state, loss_info = jax.lax.scan(
             _update_epoch, update_state, None, config["ppo_epochs"]
         )
 
-        params, opt_state, traj_batch, advantages, targets, rng = update_state
-        learner_state = LearnerState(params, opt_state, rng, env_state, last_timestep)
+        params, opt_state, _, traj_batch, advantages, targets, rng = update_state
+        learner_state = RNNLearnerState(
+            params,
+            opt_state,
+            rng,
+            env_state,
+            last_timestep,
+            last_done,
+            hstates,
+        )
         metric = traj_batch.info
         return learner_state, (metric, loss_info)
 
-    def learner_fn(learner_state: LearnerState) -> ExperimentOutput:
+    def learner_fn(learner_state: RNNLearnerState) -> ExperimentOutput:
         """Learner function.
 
         This function represents the learner, it updates the network parameters
@@ -301,6 +448,8 @@ def get_learner_fn(
                 - rng (chex.PRNGKey): The random number generator state.
                 - env_state (LogEnvState): The environment state.
                 - timesteps (TimeStep): The initial timestep in the initial trajectory.
+                - dones (bool): Whether the initial timestep was a terminal state.
+                - hstates (chex.Array): The initial hidden state of the policy and critic RNN.
         """
 
         batched_update_step = jax.vmap(_update_step, in_axes=(0, None), axis_name="batch")
@@ -323,7 +472,7 @@ def get_learner_fn(
 
 def learner_setup(
     env: Environment, rngs: chex.Array, config: Dict
-) -> Tuple[callable, ActorCritic, LearnerState]:
+) -> Tuple[Callable, ActorCritic, RNNLearnerState]:
     """Initialise learner_fn, network, optimiser, environment and states."""
     # Get available TPU cores.
     n_devices = len(jax.devices())
@@ -341,26 +490,41 @@ def learner_setup(
         optax.adam(config["lr"], eps=1e-5),
     )
 
-    # Initialise observation.
-    obs = env.observation_spec().generate_value()
-    # Select only obs for a single agent.
-    init_x = ObservationGlobalState(
-        agents_view=obs.agents_view[0],
-        action_mask=obs.action_mask[0],
-        global_state=obs.global_state,
-        step_count=obs.step_count[0],
+    # Initialise observation: Select only obs for a single agent.
+    init_obs = env.observation_spec().generate_value()
+    # init_obs = jax.tree_util.tree_map(lambda x: x[0], init_obs)
+    init_obs = jax.tree_util.tree_map(
+        lambda x: jnp.repeat(x[jnp.newaxis, ...], config["num_envs"], axis=0),
+        init_obs,
     )
-    init_x = jax.tree_util.tree_map(lambda x: x[None, ...], init_x)
+    init_obs = jax.tree_util.tree_map(lambda x: x[None, ...], init_obs)
+
+    # Select only a single agent
+    init_done = jnp.zeros((1, config["num_envs"]), dtype=bool)
+    init_obs_single = ObservationGlobalState(
+        agents_view=init_obs.agents_view[:, :, 0, :],
+        action_mask=init_obs.action_mask[:, :, 0, :],
+        global_state=init_obs.global_state,
+        step_count=init_obs.step_count[:, 0],
+    )
+
+    init_single = (init_obs_single, init_done)
+
+    # Initialise hidden state.
+    init_policy_hstate = ScannedRNN.initialize_carry((config["num_envs"]), 128)
+    init_critic_hstate = ScannedRNN.initialize_carry((config["num_envs"]), 128)
+
+    init_hiddens = (init_policy_hstate, init_critic_hstate)
 
     # initialise params and optimiser state.
-    params = network.init(rng_p, init_x)
+    params = network.init(rng_p, init_hiddens, init_single)
     opt_state = optim.init(params)
 
     # Vmap network apply function over number of agents.
     vmapped_network_apply_fn = jax.vmap(
         network.apply,
-        in_axes=(None, ObservationGlobalState(1, 1, None, 1)),
-        out_axes=(1, 1),
+        in_axes=(None, (1, None), (ObservationGlobalState(2, 2, None, 2), None)),
+        out_axes=((1, None), 2, 2),
     )
 
     # Get batched iterated update and replicate it to pmap it over cores.
@@ -371,6 +535,12 @@ def learner_setup(
     broadcast = lambda x: jnp.broadcast_to(x, (n_devices, config["update_batch_size"]) + x.shape)
     params = jax.tree_map(broadcast, params)
     opt_state = jax.tree_map(broadcast, opt_state)
+
+    # Duplicate the hidden state for each agent.
+    init_policy_hstate = jnp.expand_dims(init_policy_hstate, axis=1)
+    init_policy_hstate = jnp.tile(init_policy_hstate, (1, config["num_agents"], 1))
+    policy_hstates = jax.tree_map(broadcast, init_policy_hstate)
+    critic_hstates = jax.tree_map(broadcast, init_critic_hstate)
 
     # Initialise environment states and timesteps.
     rng, *env_rngs = jax.random.split(
@@ -391,7 +561,21 @@ def learner_setup(
     env_states = jax.tree_util.tree_map(reshape_states, env_states)
     timesteps = jax.tree_util.tree_map(reshape_states, timesteps)
 
-    init_learner_state = LearnerState(params, opt_state, step_rngs, env_states, timesteps)
+    # Initialise dones.
+    dones = jnp.zeros(
+        (
+            n_devices,
+            config["update_batch_size"],
+            config["num_envs"],
+            config["num_agents"],
+        ),
+        dtype=bool,
+    )
+    hstates = (policy_hstates, critic_hstates)
+
+    init_learner_state = RNNLearnerState(
+        params, opt_state, step_rngs, env_states, timesteps, dones, hstates
+    )
     return learn, network, init_learner_state
 
 
@@ -427,6 +611,8 @@ def run_experiment(_run: run.Run, _config: Dict, _log: SacredLogger) -> None:
         params=learner_state.params,
         config=config,
         centralised_critic=True,
+        use_recurrent_net=True,
+        scanned_rnn=ScannedRNN,
     )
 
     # Calculate total timesteps.
@@ -504,14 +690,14 @@ def hydra_entry_point(cfg: DictConfig) -> None:
     ex.captured_out_filter = utils.apply_backspaces_and_linefeeds
     results_path = os.path.join(dirname(dirname(abspath(__file__))), "results")
 
-    exp_path = get_experiment_path(cfg, "ff_mappo")
+    exp_path = get_experiment_path(cfg, "rec_mappo")
     file_obs_path = os.path.join(results_path, exp_path)
     ex.observers = [observers.FileStorageObserver.create(file_obs_path)]
     ex.add_config(OmegaConf.to_container(cfg, resolve=True))
     ex.main(run_experiment)
     ex.run(config_updates={})
 
-    print(f"{Fore.CYAN}{Style.BRIGHT}MAPPO experiment completed{Style.RESET_ALL}")
+    print(f"{Fore.CYAN}{Style.BRIGHT} Recurrent MAPPO experiment completed{Style.RESET_ALL}")
 
 
 if __name__ == "__main__":
